fix(natives): kept embed reset independent of generated manifest
- Moved the version-sentinel helpers into native/version-sentinel.js so the
  embed script no longer imports loader-state.js and its generated manifest.
- Restored gen:native:reset from an inconsistent tree (manifest present,
  archive deleted) instead of failing on the missing archive import.

Fixes #11831
