+++ packages/natives/native/loader-state.d.ts
@@ -18,12 +18,6 @@ export interface EmbeddedAddon {
 	archive?: EmbeddedAddonArchive;
 }
 
-/** Return the native-addon export expected for a package version. */
-export function versionSentinelFor(packageVersion: string): string;
-
-/** Check whether addon bytes contain the exact expected version sentinel. */
-export function containsVersionSentinel(bytes: Buffer, expected: string): boolean;
-
 export interface DetectCompiledBinaryInput {
 	embeddedAddon: EmbeddedAddon | null | undefined;
 	env: Record<string, string | undefined>;
