+++ packages/natives/native/loader-state.js
@@ -6,6 +6,7 @@ import * as path from "node:path";
 import * as zlib from "node:zlib";
 import packageJson from "../package.json" with { type: "json" };
 import { embeddedAddon } from "./embedded-addon.js";
+import { containsVersionSentinel, versionSentinelFor } from "./version-sentinel.js";
 
 /**
  * Native addon loader for `@oh-my-pi/pi-natives`.
@@ -75,37 +76,6 @@ function resolveLeafPackageDir(platformTag) {
 // =========================================================================
 // Pure helpers — re-exported for unit tests in `packages/natives/test/`.
 // =========================================================================
-
-/**
- * Return the version sentinel exported by an addon built for `packageVersion`.
- * @param {string} packageVersion
- * @returns {string}
- */
-export function versionSentinelFor(packageVersion) {
-	return `__piNativesV${packageVersion.replace(/[^A-Za-z0-9]/g, "_")}`;
-}
-
-/**
- * Check for an exact version sentinel rather than a longer sentinel with the
- * expected value as its prefix.
- * @param {Buffer} bytes
- * @param {string} expected
- * @returns {boolean}
- */
-export function containsVersionSentinel(bytes, expected) {
-	if (expected.length === 0) return false;
-	let offset = 0;
-	while (offset < bytes.length) {
-		const index = bytes.indexOf(expected, offset);
-		if (index === -1) return false;
-		const next = bytes[index + expected.length];
-		const isIdentifierByte =
-			next === 95 || (next >= 48 && next <= 57) || (next >= 65 && next <= 90) || (next >= 97 && next <= 122);
-		if (!isIdentifierByte) return true;
-		offset = index + expected.length;
-	}
-	return false;
-}
 
 /**
  * @param {{
